requirements.txt
fastapi
uvicorn
osmnx
networkx
numpy
shapely
python-dotenv